refactor: Menu 도메인을 멀티모듈로 분리한다.
