+++ menu/src/main/java/kitchenpos/menu/domain/repository/MenuProductRepository.java
@@ -1,6 +1,7 @@
-package kitchenpos.domain.repository;
+package kitchenpos.menu.domain.repository;
 
-import kitchenpos.domain.MenuProduct;
+
+import kitchenpos.menu.domain.MenuProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
