+++ menu/src/main/java/kitchenpos/menu/dto/request/MenuGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.menu.dto.request;
 
 public class MenuGroupCreateRequest {
