+++ menu/src/main/java/kitchenpos/menu/ui/MenuProductRestController.java
@@ -1,6 +1,6 @@
-package kitchenpos.ui;
+package kitchenpos.menu.ui;
 
-import kitchenpos.application.MenuProductService;
+import kitchenpos.menu.application.MenuProductService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
